requirements.txt
pandas
numpy
scikit-learn
xgboost
imbalanced-learn
joblib
shap
networkx
scipy
fastapi
uvicorn
pydantic
streamlit
plotly
requests
supabase
openai
python-dotenv